优化

+++ src/main/java/cn/offway/hades/controller/ProductController.java
@@ -209,15 +209,17 @@ public class ProductController {
 		List<PhProductInfo> phProductInfos = phProductInfoService.findAll(Arrays.asList(ids));
 		for (PhProductInfo phProductInfo : phProductInfos) {
 			if(StringUtils.isNotBlank(status)){
-				if(!phProductInfo.getStatus().equals(status)&& phProductInfo.getBeginTime().after(new Date())){
-					if(status.equals("1")){
-						Map<String, String> extras = new HashMap<>();
-						extras.put("type", "2");//0-H5,1-精选文章,2-活动
-						extras.put("id", null);
-						extras.put("url", null);
-						jPushService.createSingleSchedule(String.valueOf(phProductInfo.getId()), "0", "抽奖通知", phProductInfo.getBeginTime(), "抽奖通知", "【免费送】"+phProductInfo.getShareDesc(), extras);
-					}else if(status.equals("0")){
-						jPushService.deleteSchedule(String.valueOf(phProductInfo.getId()), "0");
+				if(BitUtil.has(phProductInfo.getChannel().intValue(), BitUtil.APP)){
+					if(!phProductInfo.getStatus().equals(status)&& phProductInfo.getBeginTime().after(new Date())){
+						if(status.equals("1")){
+							Map<String, String> extras = new HashMap<>();
+							extras.put("type", "2");//0-H5,1-精选文章,2-活动
+							extras.put("id", null);
+							extras.put("url", null);
+							jPushService.createSingleSchedule(String.valueOf(phProductInfo.getId()), "0", "抽奖通知", phProductInfo.getBeginTime(), "抽奖通知", "【免费送】"+phProductInfo.getShareDesc(), extras);
+						}else if(status.equals("0")){
+							jPushService.deleteSchedule(String.valueOf(phProductInfo.getId()), "0");
+						}
 					}
 				}
 				phProductInfo.setStatus(status);
